requirements.txt
pillow
pyqrcode
pypng
termcolor